Break modal into a separate component
