+++ js/components/post-select/modal.js
@@ -1,30 +1,18 @@
+import wp from 'wp';
 import React from 'react';
 import PropTypes from 'prop-types';
-import wp from 'wp';
 
-import PostSelectBrowse from '../../containers/post-select/browse';
-import PostSelectSelection from './selection';
-
-const { Button } = wp.components;
 const { __ } = wp.i18n;
+const { Button } = wp.components;
 
-const PostSelectModal = props => {
-	const {
-		modalTitle,
-		postType,
-		onSelect,
-		onClose,
-		onToggleSelected,
-		onMoveItemUp,
-		onMoveItemDown,
-		onChangeContentState,
-		contentState,
-		termFilters,
-		selection,
-		modalRef,
-	} = props;
-
-	return ( <div className="post-select post-select-modal">
+const Modal = ( {
+	modalTitle,
+	onClose,
+	modalRef,
+	modalContent,
+	modalToolbar,
+} ) => (
+	<div className="post-select post-select-modal">
 		<div className="media-modal-backdrop"></div>
 		<div
 			className="modal media-modal wp-core-ui"
@@ -39,77 +27,34 @@ const PostSelectModal = props => {
 				className="media-modal-close"
 				onClick={ () => onClose() }
 			>
-				<span className="media-modal-icon"><span className="screen-reader-text">{ __( 'Close media panel' ) }</span></span>
+				<span className="media-modal-icon">
+					<span className="screen-reader-text">
+						{ __( 'Close modal' ) }
+					</span>
+				</span>
 			</Button>
+
 			<div className="media-frame-title">
 				<h1 id="media-modal-title">{ modalTitle }</h1>
 			</div>
-			<div className="media-modal-content">
-				{ ( contentState === 'browse' ) && (
-					<PostSelectBrowse
-						selection={ selection }
-						postType={ postType }
-						onToggleSelected={ onToggleSelected }
-						termFilters={ termFilters }
-					/>
-				) }
-				{ ( contentState === 'selection' ) && (
-					<PostSelectSelection
-						selection={ selection }
-						postType={ postType }
-						onRemoveItem={ onToggleSelected }
-						onMoveItemUp={ onMoveItemUp }
-						onMoveItemDown={ onMoveItemDown }
-					/>
-				) }
-			</div>
+
+			<div className="media-modal-content">{ modalContent }</div>
+
 			<div className="media-frame-toolbar">
 				<div className="media-toolbar">
-					<Button
-						isPrimary={ true }
-						isLarge
-						onClick={ () => onSelect() }
-					>Select</Button>
-					{ contentState !== 'selection' && (
-						<Button
-							isPrimary={ false }
-							isLarge
-							onClick={ () => onChangeContentState( 'selection' ) }
-						>View Selected Posts</Button>
-					) }
-					{ contentState !== 'browse' && (
-						<Button
-							isPrimary={ false }
-							isLarge
-							onClick={ () => onChangeContentState( 'browse' ) }
-						>Browse posts</Button>
-					) }
+					{ modalToolbar }
 				</div>
 			</div>
 		</div>
-	</div> )
+	</div>
+);
+
+Modal.propTypes = {
+	onClose: PropTypes.func.isRequired,
+	modalTitle: PropTypes.string.isRequired,
+	modalContent: PropTypes.node.isRequired,
+	modalToolbar: PropTypes.node.isRequired,
+	modalRef: PropTypes.node,
 }
 
-PostSelectModal.defaultProps = {
-	postType: 'post',
-	modalTitle: __( 'Select a post' ),
-	contentState: 'browse',
-	selection: [],
-};
-
-PostSelectModal.propTypes = {
-	postType: PropTypes.string,
-	modalTitle: PropTypes.string,
-	onSelect: PropTypes.func.isRequired,
-	onClose: PropTypes.func.isRequired,
-	onToggleSelected: PropTypes.func.isRequired,
-	onMoveItemUp: PropTypes.func.isRequired,
-	onMoveItemDown: PropTypes.func.isRequired,
-	onChangeContentState: PropTypes.func.isRequired,
-	contentState: PropTypes.string.isRequired,
-	termFilters: PropTypes.array,
-	selection: PropTypes.array,
-	modalRef: PropTypes.func.isRequired,
-};
-
-export default PostSelectModal;
+export default Modal;
